fix(linux): make LogicalSize and LogicalRect use unsigned values

+++ kotlin-desktop-toolkit/src/main/kotlin/org/jetbrains/desktop/linux/Converters.kt
@@ -104,14 +104,14 @@ internal fun splitCsv(s: String): List<String> {
 }
 
 internal fun LogicalSize.Companion.fromNative(s: MemorySegment) = LogicalSize(
-    width = NativeLogicalSize.width(s),
-    height = NativeLogicalSize.height(s),
+    width = NativeLogicalSize.width(s).toUInt(),
+    height = NativeLogicalSize.height(s).toUInt(),
 )
 
 internal fun LogicalSize.toNative(arena: Arena): MemorySegment {
     val result = NativeLogicalSize.allocate(arena)
-    NativeLogicalSize.width(result, width)
-    NativeLogicalSize.height(result, height)
+    NativeLogicalSize.width(result, width.toInt())
+    NativeLogicalSize.height(result, height.toInt())
     return result
 }
 
@@ -129,10 +129,10 @@ internal fun LogicalPoint.toNative(arena: Arena): MemorySegment {
 
 internal fun LogicalRect.toNative(arena: Arena): MemorySegment {
     val result = NativeLogicalRect.allocate(arena)
-    NativeLogicalRect.x(result, x)
-    NativeLogicalRect.y(result, y)
-    NativeLogicalRect.width(result, width)
-    NativeLogicalRect.height(result, height)
+    NativeLogicalRect.x(result, x.toInt())
+    NativeLogicalRect.y(result, y.toInt())
+    NativeLogicalRect.width(result, width.toInt())
+    NativeLogicalRect.height(result, height.toInt())
     return result
 }
